I will probably get RSI if I have to wait for 4294967295 tiny breaks before the big break comes in.
Why is the -1 there when setting the tooltip for the amount of big breaks left? It causes the symptom described above. Try it yourself: enter the configuration dialog, cancel it and watch the tooltip.
Reason: RSITimer sets m_currentInterval to 0, and RSIWidget subtracts this with 1. Result: -1. And i18n() with plurals can't handle -1. Simple.

Summarized: Is that -1 really neccessary?


svn path=/trunk/playground/utils/rsibreak/; revision=488309