Revert "Make it compile without deprecated method"
The proper porting away from SmallIcon() is to use proper icons, not to
set pixmaps again. Revert it as a whole, so it can be fixed properly
from a cleaner state than the copy&paste of QStyle API.

Furthermore, this adds the unconditional hiding of any (including
future ones) deprecated functions in Qt, which is a big no-no.

This reverts commit 7bb92be77835c59512fffb6a64c90cc4f44bbf99.