add option to suppress rsibreak while fullscreen windows are visible
this aims at avoiding just started videos being interrupted, but it
equally applies to presentations and (fullscreen) games. as such, it
partially resolves #439601.

it would be better to track the screensaver inhibition state, but the
freedesktop api doesn't expose it, and neither does kde (gnome-session
does).
in principle, we could also track logind's power management inhibition
state (which is actually accessible), but of the apps i tested, only
chromium actually used that api. i suppose we could track that in
addition ...